[OSD-14200] remove duplicated code
