+++ controllers/accountpool/accountpool_controller.go
@@ -3,10 +3,6 @@ package accountpool
 import (
 	"context"
 	"fmt"
-	"strconv"
-
-	"gopkg.in/yaml.v2"
-
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
@@ -23,7 +19,6 @@ import (
 	"github.com/openshift/aws-account-operator/controllers/account"
 	"github.com/openshift/aws-account-operator/pkg/totalaccountwatcher"
 	"github.com/openshift/aws-account-operator/pkg/utils"
-	"github.com/openshift/aws-account-operator/test/fixtures"
 )
 
 const (
@@ -115,51 +110,9 @@ func (r *AccountPoolReconciler) Reconcile(ctx context.Context, request ctrl.Requ
 }
 
 func (r *AccountPoolReconciler) handleServiceQuotas(reqLogger logr.Logger, account *awsv1alpha1.Account) error {
-	reqLogger.Info("Loading Service Quotas")
-
-	cm, err := utils.GetOperatorConfigMap(r.Client)
+	parsedRegionalServiceQuotas, err := utils.ParseAccountPoolData(reqLogger, account.Spec.AccountPool, r.Client)
 	if err != nil {
-		reqLogger.Error(err, "failed retrieving configmap")
 		return err
-	}
-
-	accountpoolString, found := cm.Data["accountpool"]
-	if !found {
-		reqLogger.Error(fixtures.NotFound, "failed getting accountpool data from configmap")
-		return fixtures.NotFound
-	}
-
-	type Servicequotas map[string]string
-	type AccountPool struct {
-		IsDefault             bool                     `yaml:"default,omitempty"`
-		RegionedServicequotas map[string]Servicequotas `yaml:"servicequotas,omitempty"`
-	}
-
-	data := make(map[string]AccountPool)
-	err = yaml.Unmarshal([]byte(accountpoolString), &data)
-
-	if err != nil {
-		reqLogger.Error(err, "Failed to unmarshal yaml")
-		return err
-	}
-
-	var parsedRegionalServiceQuotas = make(awsv1alpha1.RegionalServiceQuotas)
-	// If the pool we've specified in the account doesn't exist, we need to error out
-	if poolData, ok := data[account.Spec.AccountPool]; !ok {
-		reqLogger.Error(fixtures.NotFound, "Accountpool not found")
-		return fixtures.NotFound
-	} else {
-		// for each service quota in a given region, we'll need to parse and save to use in the account spec.
-		for regionName, serviceQuota := range poolData.RegionedServicequotas {
-			var parsedServiceQuotas = make(awsv1alpha1.AccountServiceQuota)
-			for quotaCode, quotaValue := range serviceQuota {
-				qv, _ := strconv.Atoi(quotaValue)
-				parsedServiceQuotas[awsv1alpha1.SupportedServiceQuotas(quotaCode)] = &awsv1alpha1.ServiceQuotaStatus{
-					Value: qv,
-				}
-			}
-			parsedRegionalServiceQuotas[regionName] = parsedServiceQuotas
-		}
 	}
 	reqLogger.Info("Loaded Service Quotas")
